requirements.txt
PyQt6==6.6.1
bcrypt==4.1.2
SQLAlchemy==2.0.25
python-dotenv==1.0.0
pyfingerprint==1.5
reportlab==4.0.9
pandas==2.1.4
cryptography==41.0.7
pytest==7.4.3
selenium==4.16.0
pillow==10.2.0
qrcode==7.4.2
python-dateutil==2.8.2
